Refactor authentication and user management components: Updated middleware for role-based access control, replaced obsolete user and task management routes, and enhanced session handling. Refactored components to utilize new user model and improved type safety across the application.

+++ app/api/(router)/(admin_permission)/deleteTask/route.ts
@@ -11,6 +11,15 @@ export async function DELETE(req: Request) {
     if (!task) {
         return NextResponse.json({ error: "Task not found" }, { status: 404 });
     }
+    const user = await prisma.users.findUnique({
+        where: { id: task.userid },
+    });
+    if(!user) {
+        return NextResponse.json({ error: "Supplier not found" }, { status: 404 });
+    }
+    if(user.role !== "ADMIN") {
+        return NextResponse.json({ error: "Supplier is not a admin" }, { status: 400 });
+    }
     await prisma.tasks.delete({
                 where: { id },
             });
